Bug 1674777 part 4 - Change some JSAPI byteLength and byteOffset accessors to return size_t instead of uint32_t. r=sfink
Of these four, only JS_GetTypedArrayByteLength is used outside jsapi-tests.

Differential Revision: https://phabricator.services.mozilla.com/D103277

UltraBlame original commit: 94684c2cfbc4e28e82c2b6a8721d285f9a2f5bba

+++ js/public/experimental/TypedData.h
@@ -309,7 +309,7 @@ extern JS_FRIEND_API uint32_t JS_GetTypedArrayLength(JSObject* obj);
 
 
 
-extern JS_FRIEND_API uint32_t JS_GetTypedArrayByteOffset(JSObject* obj);
+extern JS_FRIEND_API size_t JS_GetTypedArrayByteOffset(JSObject* obj);
 
 
 
@@ -318,17 +318,17 @@ extern JS_FRIEND_API uint32_t JS_GetTypedArrayByteOffset(JSObject* obj);
 
 
 
-extern JS_FRIEND_API uint32_t JS_GetTypedArrayByteLength(JSObject* obj);
+extern JS_FRIEND_API size_t JS_GetTypedArrayByteLength(JSObject* obj);
 
 
 
 
-extern JS_FRIEND_API uint32_t JS_GetArrayBufferViewByteLength(JSObject* obj);
+extern JS_FRIEND_API size_t JS_GetArrayBufferViewByteLength(JSObject* obj);
 
 
 
 
-extern JS_FRIEND_API uint32_t JS_GetArrayBufferViewByteOffset(JSObject* obj);
+extern JS_FRIEND_API size_t JS_GetArrayBufferViewByteOffset(JSObject* obj);
